pyproject.toml
[tool.black]
target-version = ["py313"]

[tool.poetry]
name = "trackmate"
version = "1.0.0"
description = "Tracker and Task Manager"
authors = ["Maksim <mksmvnv@icloud.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
pip = "^24.2"
python = "^3.13"
django = "^5.1"
python-dotenv = "^1.0.1"
pillow = "^10.4.0"
djangorestframework = "^3.15.2"
django-enumfield = "^3.1"
psycopg2-binary = "^2.9.10"

[tool.poetry.dev-dependencies]
black = "^24.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
